Created addresses JSON file
The addresses JSON file will contain a list with all of the addresses that exist in Magic Cow Games' services, such as the WDW web server's address and port.

This is a patchy solution that is used to allow using github's relatively stable domain as a way to easily access MCG's services at any time without modifying game code whenever the server's real address changes.

This solution is temporary exists for now since MCG does not own a domain.

The file contains a default entry for now.